* Improved slow control. * Data flow monitor over NSM.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@7854 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ daq/rfarm/manager/src/RFDqmServer.cc
@@ -58,10 +58,12 @@ int RFDqmServer::Configure(NSMmsg*, NSMcontext*)
 int RFDqmServer::UnConfigure(NSMmsg*, NSMcontext*)
 {
   //  system("killall hrelay hserver");
-  int status;
-  kill(m_pid_dqm, SIGINT);
-  waitpid(m_pid_dqm, &status, 0);
-  printf("Unconfigre : done\n");
+  if (m_pid_dqm != 0) {
+    int status;
+    kill(m_pid_dqm, SIGINT);
+    waitpid(m_pid_dqm, &status, 0);
+    printf("Unconfigre : done\n");
+  }
   return 0;
 }
